requirements.txt
streamlit
langchain-groq
pymupdf
python-dotenv
apify-client
fastmcp